+++ frontend/src/components/Dialog/Event/Input/EventMultInputs.jsx
@@ -19,7 +19,7 @@ const EventMultInputs = ({ formik }) => {
 						value: formik.values.operation.value,
 						name: formik.values.operation.name,
 						to_bucket: formik.values.operation.to_bucket,
-						amount: parseInt(e.target.value),
+						amount: parseFloat(e.target.value),
 					});
 				}}
 				value={formik.values.operation.amount ?? 0}
